requirements.txt
pyyaml
python-ffmpeg
streamlink
pyee
